+++ packages/eslint-plugin-template/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 20.5.0 (2025-10-27)
+
+### 🚀 Features
+
+- **eslint-plugin-template:** add prefer-built-in-pipes rule ([#2605](https://github.com/angular-eslint/angular-eslint/pull/2605))
+
+### ❤️ Thank You
+
+- Andreas Nicolaou @andreasnicolaou
+
 ## 20.4.0 (2025-10-12)
 
 ### 🚀 Features